[refact] owned, 가게 권한 api 호출 횟수 감소시킴
